# Switched DataStorage from double to float (BLE compatible) # use now one BLECharacteristic to send data # Data will now be notified on BLE, no polling necessary on MobilePhone # connectionless BLE will transfer data. Without connection every 30 seconds an update elapsed, with connection every 2,5 seconds. Will be increased in future for realtime

+++ platformio-project/lib/ADSDataCollector/ADSDataCollector.cpp
@@ -54,7 +54,7 @@ void ADSDataCollector::updateData()
     #ifdef ADSDATACOLLECTOR_H_DEBUG
         Serial.println("DEBUG CODE ACTIVE! RANDOM DATA");
         long randomVal = random(100,160);
-        double result = randomVal/10;
+        float result = randomVal/10;
     #else
         float reference_vcc = 3.3; // Volt... ESP32 system voltage
         float vcc_max = 16.0; // Volt... maximum input on voltage divider
@@ -69,7 +69,7 @@ void ADSDataCollector::updateData()
         float voutRaw = vout / (reference_vcc / vcc_max);
         float rFactor = (float)resistor_to_gnd_max / (float)resistor_to_gnd; // It is important to cast a float!!
 
-        double result =  (voutRaw * rFactor);
+        float result =  (voutRaw * rFactor);
     #endif
     
     _dataStorage->addData(_channelName + CHANNELEXTVCC , result);
