+++ platformio-project/lib/BMEDataCollector/BMEDataCollector.h
@@ -22,9 +22,9 @@
 class BMEDataCollector : public IDataCollector {
        public:
               BMEDataCollector(std::string channelName);
-              BMEDataCollector(std::string channelName, double tempoeratureOffset);
+              BMEDataCollector(std::string channelName, float tempoeratureOffset);
               BMEDataCollector(std::string channelName, byte address,byte i2cSDA,byte i2cSCL);
-              BMEDataCollector(std::string channelName, byte address,byte i2cSDA,byte i2cSCL, double tempoeratureOffset);
+              BMEDataCollector(std::string channelName, byte address,byte i2cSDA,byte i2cSCL, float tempoeratureOffset);
               void init(IDataStorage *storage);
               void reInit();
               void sleep();
@@ -35,9 +35,9 @@ class BMEDataCollector : public IDataCollector {
               BME280_I2C *_bme280; //BME280_I2C bme(0x76); // uint i2c-address
               std::string _channelName = "BMECollector";
               void readBMEData();     
-              double _temperatureOffset=0;
-              double _temp = 0;
-              double _pressure = 0;    
+              float _temperatureOffset=0;
+              float _temp = 0;
+              float _pressure = 0;    
               bool _bmeIsReady=0;
               byte _i2cSDA=GPIO_NUM_21;
               byte _i2cSCL=GPIO_NUM_22;
