+++ platformio-project/lib/DTSDataCollector/DTSDataCollector.h
@@ -25,7 +25,7 @@ class DTSDataCollector : public IDataCollector {
        public:
               DTSDataCollector(std::string channelName);
               DTSDataCollector(std::string channelName,uint8_t oneWirePin);
-              DTSDataCollector(std::string channelName, double temperatureOffset,uint8_t oneWirePin = 5);
+              DTSDataCollector(std::string channelName, float temperatureOffset,uint8_t oneWirePin = 5);
               void init(IDataStorage *storage);
               void reInit();
               void sleep();
@@ -37,9 +37,9 @@ class DTSDataCollector : public IDataCollector {
 
               std::string _channelName = "DTSCollector";
               uint8_t _oneWirePin = GPIO_NUM_23;
-              double _temperatureOffset = 0;
-              double _temp = 0;
-              double _pressure = 0;    
+              float _temperatureOffset = 0;
+              float _temp = 0;
+              float _pressure = 0;    
 };
 
 #endif
